Split stats
